veteran-ai/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & Vector DB
supabase==2.0.2
psycopg2-binary==2.9.9
pgvector==0.2.4

# Data Processing
pandas==2.1.4
numpy==1.24.4

# Embedding & LLM
openai==1.51.0
anthropic==0.34.2
sentence-transformers==2.2.2

# Document Processing
pytesseract==0.3.10
pdf2image==1.16.3
PyPDF2==3.0.1
python-docx==1.1.0

# Audio Processing  
openai-whisper==20231117
pydub==0.25.1

# Chat API Integration
slack-sdk==3.26.1
slack-bolt==1.18.1
requests==2.31.0

# Web Framework
jinja2==3.1.2
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
pillow==10.1.0
aiofiles==23.2.1
celery==5.3.4
redis==5.0.1

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0